requirements.txt
fastapi
uvicorn
python-dotenv
pyyaml
langchain
langchain-community
langchain-openai
langchain-google-genai
langchain-groq
langchain-mistralai
langgraph
pydantic
pandas
openpyxl
python-multipart
pytest
httpx
